requirements.txt
argparse==1.2.1
click==6.3
cm-api==11.0.0
datadiff==2.0.0
decorator==4.0.6
ecdsa==0.13
Jinja2==2.8
MarkupSafe==0.23
networkx==1.11
paramiko== 1.16.0
pycrypto==2.6.1
pyyaml==3.11
requests==2.9.1
